--- web/html/src/manager/maintenance/index.js
@@ -1,4 +1,0 @@
-export default {
-  'maintenance/maintenance-windows': () => import('./maintenance-windows'),
-  'maintenance/system-assignment': () => import('./ssm/system-assignment')
-}
