+++ web/html/src/manager/maintenance/index.ts
@@ -1,0 +1,4 @@
+export default {
+  "maintenance/maintenance-windows": () => import("./maintenance-windows"),
+  "maintenance/system-assignment": () => import("./ssm/system-assignment"),
+};
